Adapt repository cloning to docker requirements
This changes how the actual PostBOUND repository is cloned. More
specifically, the repo is now cloned into a temporary directory and
moved to the actual target dir afterwards. This change was necessary,
because the PostBOUND pipeline creates all result files in a directory
under the PostBOUND root. However, these results should be shared with
the Docker client and not be restricted to the Docker container.
Therefore, the same directory (wich obviously does not exist yet when
the Docker container is created) is bound to a directory on the local
file system. This means, that the container directory is created during
container/image creation. However, this causes Git to refuse cloning
into an already existing directory. Therefore, a temporary dir was used
and later "merged" by moving into the actual target directory.